WIP: create timer_helper; refactor side dock
